moved UpdateEventDetail to components
