+++ frontend/src/pages/UpdateEvent.jsx
@@ -1,13 +1,8 @@
 import { React, useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import {
-  Button, Form, DatePicker, Input,
-} from 'antd';
-
-const { RangePicker } = DatePicker;
 import Interviewers from '../components/Interviewers';
-import Interviewees from '../components/Interviewees';
+import EventDetails from '../components/UpdateEventDetails';
 
 const UpdateEvent = ({ user }) => {
   const { eventId } = useParams();
@@ -35,121 +30,6 @@ const UpdateEvent = ({ user }) => {
       <Interviewers users={eventInfo.interviewers} />
     </div>
   );
-};
-
-const EventDetails = (props) => {
-  const { eventInfo, eventId } = props;
-  const onFinish = (values) => {
-    fetch(`${process.env.REACT_APP_SERVER_URL}/events/${eventId}/update`,
-      {
-        method: 'PUT',
-        credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          title: values.title,
-          description: values.description,
-          startDate: values.daterange[0],
-          endDate: values.daterange[1],
-          interviewersNeeded: values.interviewersNeeded,
-        }),
-      });
-  };
-  return (
-    <>
-      <h3>{eventInfo.title}</h3>
-      <h3>{eventInfo.startDate}</h3>
-      <h3>{eventInfo.endDate}</h3>
-      <h3>{eventInfo.description}</h3>
-      <h3>{eventInfo.interviewersNeeded}</h3>
-      <h3>{eventId}</h3>
-      <Form onFinish={onFinish}>
-        <Form.Item
-          label='Event Title'
-          name='title'
-        >
-          <Input type='text' />
-        </Form.Item>
-        <Form.Item
-          label='Start And End Dates'
-          name='daterange'
-        >
-          <RangePicker
-            showTime={{ format: 'HH:mm' }}
-            format='YYYY-MM-DD HH:mm'
-          />
-        </Form.Item>
-        <Form.Item
-          label='Description'
-          name='description'
-        >
-          <Input type='text' />
-        </Form.Item>
-        <Form.Item
-          label='Interviewers Needed'
-          name='interviewersNeeded'
-        >
-          <Input type='number' />
-        </Form.Item>
-        <Form.Item>
-          <Button type='primary' htmlType='submit'>
-            Update Event
-          </Button>
-        </Form.Item>
-      </Form>
-    </>
-  );
-};
-
-const Interviewers = ({ users }) => (
-  <>
-    <div>Current Interviewers:</div>
-    {
-      users.map((i) => <Interviewer key={i} userId={i.userId} />)
-    }
-    <Button>
-      Add interviewers
-    </Button>
-  </>
-);
-
-const Interviewer = ({ userId }) => {
-  const [userInfo, updateUserInfo] = useState({});
-
-  useEffect(() => {
-    fetch(`${process.env.REACT_APP_SERVER_URL}/users/${userId}`,
-      { credentials: 'include' }).then((res) => res.json())
-      .then((user) => {
-        updateUserInfo(user);
-      });
-  }, [userId]);
-
-  return (
-    <div>
-      {userInfo.name}
-      (
-      {userInfo.email}
-      )
-      eventId:
-      {eventId}
-      <br />
-      event:
-      {eventInfo.title}
-      <br />
-      <Interviewers users={eventInfo.interviewers} updateEventInfo={updateEventInfo} />
-      <Interviewees users={eventInfo.interviewees} updateEventInfo={updateEventInfo} />
-    </div>
-  );
-};
-
-EventDetails.propTypes = {
-  eventInfo: PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    startDate: PropTypes.string.isRequired,
-    endDate: PropTypes.string.isRequired,
-    description: PropTypes.string,
-    interviewersNeeded: PropTypes.number.isRequired,
-  }).isRequired,
-  eventId: PropTypes.string.isRequired,
 };
 
 UpdateEvent.propTypes = {
